asdf and ginkgo updates to bring more in line with terra (#256)
* Brings things more in sync with terra.
I could not figure out how to get the gauntlet.yml, and the e2e yml github action jobs to run nix without have an hour startup for the first command on every run. Tried using caches and such in different ways to no avail. It does just seem to work for the relay.yml so I left that using nix.
I did fix up the gauntlet and e2e ymls so all we need to do is drop in the nix install and prepend the nix develop call so it should be an easy update once we get nix issues figured out.
I left us at go 1.17.x for now even though 1.18 is out since we probably don't want to add that extra complexity to testing this close to release

* Update ci job names to be more explicit for what they are doing

* for now run on every commit to it can be made a required check

* updates from comments and make script to go from tool-versions to env

* copy paste error

* looks like the importer makes everything lower case

* move to setup go v2

* pr comment fixes

* merge conflict fixes

* remove nix from relay in ci