Implement Basic GetLatestValue for Solana
This commit fills in the functionality for `GetLatestValue` such that
one or more accounts with an associated Anchor IDl can be read and
mapped to native Go structs. Methods and prodecures can be defined in
a config where an IDL and codec modifiers can be provided.

Current limitations include completing enums and defining max length
of strings.