Make RegisterFilter operation atomic
Changing the behavior of addToIndices to ignore db corruption when detected, instead of aborting the operation is slightly better behavior for
the case where it's called by RegisterFilter. Because the filter passed to addToIndices was just successfully inserted into the db, updating
the in-memory index to match may fix the corruption issue whereas ignoring it will simply report the problem and leave things in an inconsistent
state.

addToIndices is also called from LoadFilters(). In that context, whether the operation is aborted or not makes no difference: the error happens
if 2 rows with the same name or filter_id have just been read from the filters table. Aborting amounts to setting the in-memory state to match
the first row, while ignoring sets the in-memory state to match the second row. We could go a step further and try to fix the inconsistency
by deleting one of the duplicate rows and keeping the other, but instead we simply log it as a CRITICAL error to avoid erasing potentially important
information which could help someone investigate how the db got into this invalid state.

If we think of the db as the "source of truth" and the in-memory indices as a cache of it, then both of these decisions can be summarized with:
always make sure that the in-memory state matches the latest known db state as closely as possible, rather than changing the db to match what's
in memory.