Add support for lookback window
Uses the maximum of any non-permanent retention periods defined on any filters,
or 30 days by default if all filters are configured for permanent retention.

This feature allows LogPoller to skip ahead to the beginning of the lookback window
if the node has been down for a while, rather than processing every block.

GetFirstAvailableBlock optimization: if this returns something more
recent than the computed start of the lookback window, it will be used
instead.